chore(zero-cache): Move version string to types
It is going to be used for Postgres too so it should not be in paths.
